requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.1
django-bootstrap4==22.3
django-chartjs==2.3.0
django-stdimage==6.0.1
Pillow==9.2.0
pytz==2022.2.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
